requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cryptography==37.0.4
defusedxml==0.7.1
Flask==2.2.2
Flask-MySQLdb==1.0.1
greenlet==1.1.2
h11==0.13.0
httpie @ https://github.com/jakubroztocil/httpie/archive/master.tar.gz
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
multidict==6.0.2
mysqlclient==2.1.1
numpy==1.23.2
outcome==1.2.0
pandas==1.4.3
pycparser==2.21
Pygments==2.13.0
PyMySQL==1.0.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.5.1
scikit-learn==1.1.2
scipy==1.9.0
selenium==4.4.2
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.40
threadpoolctl==3.1.0
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.11
webdriver-manager==3.8.3
Werkzeug==2.2.2
wincertstore==0.2
wsproto==1.1.0
zipp==3.8.1
